calculate site-packages location portably
Uninstall virtualenv's pip in a way that works on Windows, i.e. not
using the executable that's being uninstalled and not using the code
that comes with virtualenv, which is broken.

Add '-' to Path for relative path calculation

Silence virtualenv preparation

Encourage earlier GC of the last TestPipEnvironment, which was
happening too late